Projeto Bê-á-Bá/Projeto-Docker/python/requirements.txt
flask
pandas
google-cloud-storage
psycopg2-binary
```